Reorganize Scripting into separate files; add headers; remove unused junk
