+++ hybrasyl/Scripting/HybrasylMap.cs
@@ -1,0 +1,42 @@
+﻿/*
+ * This file is part of Project Hybrasyl.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the Affero General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * without ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the Affero General Public License
+ * for more details.
+ *
+ * You should have received a copy of the Affero General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * (C) 2013 Justin Baugh (baughj@hybrasyl.com)
+ * (C) 2015-2016 Project Hybrasyl (info@hybrasyl.com)
+ *
+ * For contributors and individual authors please refer to CONTRIBUTORS.MD.
+ * 
+ */
+
+namespace Hybrasyl.Scripting
+{
+
+    public class HybrasylMap
+    {
+        private Map Map { get; set; }
+
+        public HybrasylMap(Map map)
+        {
+            Map = map;
+        }
+
+        public bool DropItem(string name, int x = -1, int y = -1)
+        {
+            return false;
+        }
+
+    }
+
+}
