+++ hybrasyl/Scripting/HybrasylWorldObject.cs
@@ -1,0 +1,93 @@
+﻿/*
+ * This file is part of Project Hybrasyl.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the Affero General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * without ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the Affero General Public License
+ * for more details.
+ *
+ * You should have received a copy of the Affero General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * (C) 2013 Justin Baugh (baughj@hybrasyl.com)
+ * (C) 2015-2016 Project Hybrasyl (info@hybrasyl.com)
+ *
+ * For contributors and individual authors please refer to CONTRIBUTORS.MD.
+ * 
+ */
+
+
+using Hybrasyl.Objects;
+
+namespace Hybrasyl.Scripting
+{
+    public class HybrasylWorldObject
+    {
+        internal WorldObject Obj { get; set; }
+
+        public HybrasylWorldObject(WorldObject obj)
+        {
+            Obj = obj;
+        }
+
+        public void DisplayPursuits(dynamic invoker)
+        {
+            if (Obj is Merchant)
+            {
+                var merchant = Obj as Merchant;
+                if (invoker is HybrasylUser)
+                {
+                    var hybUser = (HybrasylUser)invoker;
+                    merchant.DisplayPursuits(hybUser.User);
+                }
+            }
+
+        }
+
+        public void Destroy()
+        {
+            if (Obj is ItemObject || Obj is Gold)
+            {
+                Game.World.Remove(Obj);
+            }
+        }
+
+        public void AddPursuit(HybrasylDialogSequence hybrasylSequence)
+        {
+            if (Obj is VisibleObject && !(Obj is User))
+            {
+                var vobj = Obj as VisibleObject;
+                vobj.AddPursuit(hybrasylSequence.Sequence);
+            }
+
+        }
+
+        public void RegisterSequence(HybrasylDialogSequence hybrasylSequence)
+        {
+            if (Obj is VisibleObject && !(Obj is User))
+            {
+                var vobj = Obj as VisibleObject;
+                vobj.RegisterDialogSequence(hybrasylSequence.Sequence);
+            }
+        }
+
+        public void RegisterGlobalSequence(HybrasylDialogSequence globalSequence)
+        {
+            Game.World.RegisterGlobalSequence(globalSequence.Sequence);
+        }
+
+        public void Say(string message)
+        {
+            if (Obj is Creature)
+            {
+                var creature = Obj as Creature;
+                creature.Say(message);
+            }
+        }
+
+    }
+}
