+++ hybrasyl/Scripting/ScriptInvocation.cs
@@ -1,0 +1,50 @@
+﻿/*
+ * This file is part of Project Hybrasyl.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the Affero General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * without ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the Affero General Public License
+ * for more details.
+ *
+ * You should have received a copy of the Affero General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * (C) 2013 Justin Baugh (baughj@hybrasyl.com)
+ * (C) 2015-2016 Project Hybrasyl (info@hybrasyl.com)
+ *
+ * For contributors and individual authors please refer to CONTRIBUTORS.MD.
+ * 
+ */
+ 
+ using Hybrasyl.Objects;
+
+namespace Hybrasyl.Scripting
+{
+    public struct ScriptInvocation
+    {
+        public dynamic Function;
+        public Script Script;
+        public WorldObject Associate;
+        public WorldObject Invoker;
+
+        public ScriptInvocation(dynamic function = null, WorldObject associate = null, WorldObject invoker = null, Script script = null)
+        {
+            Function = function;
+            Associate = associate;
+            Invoker = invoker;
+            Script = null;
+        }
+
+        public bool Execute(params object[] parameters)
+        {
+            if (Script != null)
+                return Script.ExecuteFunction(this, parameters);
+            else
+                return Associate.Script.ExecuteFunction(this, parameters);
+        }
+    }
+}
